requirements.txt
asgiref==3.8.1
certifi==2025.1.31
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
logging==0.4.9.6
psycopg2==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
